Script nodes and trees for each npc added.
