+++ Ladybug Proof of Concept/Assets/Scenes/Personal Scenes/AllyScene1.0.unity.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: 21b4e33dcd4a2d8478ffe4ee3a5ed365
-folderAsset: yes
+guid: bd42006441dceeb4ea99757e90314983
 DefaultImporter:
   externalObjects: {}
   userData:
