+++ Ladybug Proof of Concept/Assets/Scripts/Duel Assets/AgrEnemy.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 924b5bbcccc2c41cf8390a528fe364b9
+guid: 478937d1bf070724998d9f87ff10b1c9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
