+++ Ladybug Proof of Concept/Assets/Scripts/Duel Assets/DuelManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e4227f615db5341bcac1c59fd102c46b
+guid: 54d071e0fc5e65542be446994fa40d04
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
