+++ Ladybug Proof of Concept/Assets/Scripts/Vera Scripts/DialogueManager.cs.meta
