+++ Ladybug Proof of Concept/Assets/Scripts/Vera Scripts/NPC.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class NPC : MonoBehaviour {
+
+    public string talkToNode;
+    public DialogueTree treeToLoad;
+}
