+++ Ladybug Proof of Concept/Assets/Scripts/Vera Scripts/NPC.cs.meta
