+++ Ladybug Proof of Concept/Assets/Vera Scriptable Objects/AInitNode1.asset
@@ -1,0 +1,24 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 606dab1dcf00c4741bf61375d0db1af1, type: 3}
+  m_Name: AInitNode1
+  m_EditorClassIdentifier: 
+  nodeName: PersonANode1
+  dialogueLines:
+  - line: Whaddya want?
+    valuableInfo: 0
+    duelTrigger: 0
+  - line: "I\u2019m busy, go away!"
+    valuableInfo: 0
+    duelTrigger: 1
+  - line: "I ain\u2019t tellin you anything, you\u2019ve got no determination."
+    valuableInfo: 0
+    duelTrigger: 0
