+++ Ladybug Proof of Concept/Assets/Vera Scriptable Objects/AInitNode1.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 21b4e33dcd4a2d8478ffe4ee3a5ed365
-folderAsset: yes
-DefaultImporter:
+guid: c0744cbfad995ed499ae3effa0ee91d4
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
