+++ Ladybug Proof of Concept/Assets/Vera Scriptable Objects/ASecondNode1.asset
@@ -1,0 +1,24 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 606dab1dcf00c4741bf61375d0db1af1, type: 3}
+  m_Name: ASecondNode1
+  m_EditorClassIdentifier: 
+  nodeName: PersonANode2
+  dialogueLines:
+  - line: "Why\u2019d you go and freak out my sister you goddamn jerk?!"
+    valuableInfo: 0
+    duelTrigger: 0
+  - line: "She didn\u2019t do anything, what\u2019s wrong with you?!"
+    valuableInfo: 0
+    duelTrigger: 1
+  - line: "FUCK OFF, SISTER-BULLY OR I\u201DLL BEAT YOUR ASS!"
+    valuableInfo: 0
+    duelTrigger: 0
