+++ Ladybug Proof of Concept/Assets/Vera Scriptable Objects/ASecondNode1.asset.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 08da9b54867e048c58ea6641f3ea0655
-DefaultImporter:
+guid: 8e062c0454db99d4aaaeab9ddeb365ff
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
