+++ Ladybug Proof of Concept/Assets/Vera Scriptable Objects/AThirdNode1.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 21b4e33dcd4a2d8478ffe4ee3a5ed365
-folderAsset: yes
-DefaultImporter:
+guid: b30df789464134b4cb68110c3a2c3ff6
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
