+++ Ladybug Proof of Concept/Assets/Vera Scriptable Objects/AllyTree.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 075334b926d654961be97b2678c6f974, type: 3}
+  m_Name: AllyTree
+  m_EditorClassIdentifier: 
+  treeName: PersonANodes
+  dialogueNodes:
+  - {fileID: 11400000, guid: c0744cbfad995ed499ae3effa0ee91d4, type: 2}
+  - {fileID: 11400000, guid: 8e062c0454db99d4aaaeab9ddeb365ff, type: 2}
+  - {fileID: 11400000, guid: b30df789464134b4cb68110c3a2c3ff6, type: 2}
