+++ Ladybug Proof of Concept/Assets/Vera Scriptable Objects/AllyTree2.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 075334b926d654961be97b2678c6f974, type: 3}
+  m_Name: AllyTree2
+  m_EditorClassIdentifier: 
+  treeName: PersonBNodes
+  dialogueNodes:
+  - {fileID: 11400000, guid: 530c08fa2cd4d5740a8bb97100ce4cec, type: 2}
+  - {fileID: 11400000, guid: 8402589d28aaae842b63ed98231f8c1b, type: 2}
+  - {fileID: 11400000, guid: cf348a6734cfe204fbcedece0cdc071a, type: 2}
