+++ Ladybug Proof of Concept/Assets/Vera Scriptable Objects/AllyTree2.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 21b4e33dcd4a2d8478ffe4ee3a5ed365
-folderAsset: yes
-DefaultImporter:
+guid: d1b1d3b2b40e99e4d96ca447a18a9721
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
