+++ Ladybug Proof of Concept/Assets/Vera Scriptable Objects/BSecondNode1.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 21b4e33dcd4a2d8478ffe4ee3a5ed365
-folderAsset: yes
-DefaultImporter:
+guid: 8402589d28aaae842b63ed98231f8c1b
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
