+++ Ladybug Proof of Concept/Assets/Vera Scriptable Objects/BThirdNode1.asset
@@ -1,0 +1,28 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 606dab1dcf00c4741bf61375d0db1af1, type: 3}
+  m_Name: BThirdNode1
+  m_EditorClassIdentifier: 
+  nodeName: PersonBNode3
+  dialogueLines:
+  - line: You wanna know where we were?
+    valuableInfo: 0
+    duelTrigger: 0
+  - line: "Um, let me think for a second. I think I\u2019m still in shock from this
+      whole thing."
+    valuableInfo: 0
+    duelTrigger: 0
+  - line: "During cocktail hour...um\u2026"
+    valuableInfo: 0
+    duelTrigger: 1
+  - line: "Listen, I really can\u2019t tell you anything else. Sorry."
+    valuableInfo: 0
+    duelTrigger: 0
