--- Ladybug Proof of Concept/Assets/Vera Scripts/NPC.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class NPC : MonoBehaviour {
-
-    public string talkToNode;
-    public DialogueTree treeToLoad;
-}
